Moved failure_snack_bar.dart to under utilities/
