+++ packages/clima_ui/lib/utilities/failure_snack_bar.dart
